Implemented search on small screens now
