+++ common/compose/src/commonMain/kotlin/app/campfire/common/compose/navigation/LocalSearchView.kt
@@ -1,0 +1,14 @@
+package app.campfire.common.compose.navigation
+
+import androidx.compose.runtime.compositionLocalOf
+
+val LocalSearchView = compositionLocalOf<SearchViewNavigationState?> { null }
+
+class SearchViewNavigationState(
+  private val onNavigate: () -> Unit,
+) {
+
+  fun navigateToSearchView() {
+    onNavigate()
+  }
+}
