+++ features/search/ui/src/commonMain/kotlin/app/campfire/search/ui/CampfireDockedSearchBar.kt
@@ -3,94 +3,46 @@ package app.campfire.search.ui
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.FlowRow
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.GridItemSpan
-import androidx.compose.foundation.lazy.grid.LazyGridScope
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.Clear
 import androidx.compose.material.icons.rounded.Search
-import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.DockedSearchBar
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.SearchBarDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.withStyle
-import androidx.compose.ui.unit.dp
 import app.campfire.common.compose.di.rememberComponent
-import app.campfire.common.compose.layout.ContentLayout
-import app.campfire.common.compose.layout.LocalContentLayout
-import app.campfire.common.compose.widgets.AuthorCard
-import app.campfire.common.compose.widgets.EmptyState
-import app.campfire.common.compose.widgets.ItemCollectionCard
-import app.campfire.common.compose.widgets.LibraryItemCard
-import app.campfire.core.di.UserScope
-import app.campfire.core.model.Author
-import app.campfire.core.model.BasicSearchResult
-import app.campfire.core.model.LibraryItem
-import app.campfire.core.model.Series
-import app.campfire.search.api.SearchResult
-import app.campfire.search.ui.composables.BasicSearchResultChip
+import app.campfire.search.ui.composables.SearchResultContent
+import app.campfire.search.ui.di.SearchUiComponent
 import campfire.features.search.ui.generated.resources.Res
-import campfire.features.search.ui.generated.resources.header_authors
-import campfire.features.search.ui.generated.resources.header_books
-import campfire.features.search.ui.generated.resources.header_genres
-import campfire.features.search.ui.generated.resources.header_narrators
-import campfire.features.search.ui.generated.resources.header_series
-import campfire.features.search.ui.generated.resources.header_tags
-import campfire.features.search.ui.generated.resources.search_empty_message
 import campfire.features.search.ui.generated.resources.search_placeholder_format
 import campfire.features.search.ui.generated.resources.search_placeholder_options
-import campfire.features.search.ui.generated.resources.search_results_empty_message
-import campfire.features.search.ui.generated.resources.search_results_error_message
-import com.r0adkll.kimchi.annotations.ContributesTo
 import com.slack.circuit.runtime.Navigator
-import kotlin.contracts.ExperimentalContracts
-import kotlin.contracts.InvocationKind
-import kotlin.contracts.contract
 import org.jetbrains.compose.resources.stringArrayResource
 import org.jetbrains.compose.resources.stringResource
-
-@ContributesTo(UserScope::class)
-interface DockedSearchBarComponent {
-  val searchPresenterFactory: SearchPresenterFactory
-}
 
 @Composable
 fun CampfireDockedSearchBar(
   navigator: Navigator,
   modifier: Modifier = Modifier,
-  component: DockedSearchBarComponent = rememberComponent(),
+  component: SearchUiComponent = rememberComponent(),
 ) {
-  val presenter = remember(component, navigator) { component.searchPresenterFactory(navigator) }
+  var expanded by remember { mutableStateOf(false) }
+
+  val presenter = remember(component, navigator) { component.searchPresenterFactory(navigator) { expanded = false } }
   val uiState = presenter.present()
+
   CampfireDockedSearchBar(
     state = uiState,
+    expanded = expanded,
+    onExpandedChange = { expanded = it },
     modifier = modifier,
   )
 }
@@ -98,11 +50,12 @@ fun CampfireDockedSearchBar(
 @Composable
 private fun CampfireDockedSearchBar(
   state: SearchUiState,
+  expanded: Boolean,
+  onExpandedChange: (Boolean) -> Unit,
   modifier: Modifier = Modifier,
 ) {
   val eventSink = state.eventSink
 
-  var expanded by remember { mutableStateOf(false) }
   DockedSearchBar(
     inputField = {
       SearchBarDefaults.InputField(
@@ -114,9 +67,7 @@ private fun CampfireDockedSearchBar(
           eventSink(SearchUiEvent.QueryChanged(it))
         },
         expanded = expanded,
-        onExpandedChange = {
-          expanded = it
-        },
+        onExpandedChange = onExpandedChange,
         leadingIcon = {
           Icon(Icons.Rounded.Search, contentDescription = null)
         },
@@ -129,7 +80,8 @@ private fun CampfireDockedSearchBar(
             IconButton(
               onClick = {
                 eventSink(SearchUiEvent.ClearQuery)
-                expanded = false
+                eventSink(SearchUiEvent.Dismiss)
+                onExpandedChange(false)
               },
             ) {
               Icon(Icons.Rounded.Clear, contentDescription = null)
@@ -148,7 +100,7 @@ private fun CampfireDockedSearchBar(
       )
     },
     expanded = expanded,
-    onExpandedChange = { expanded = it },
+    onExpandedChange = onExpandedChange,
     modifier = modifier,
   ) {
     SearchResultContent(
@@ -162,253 +114,4 @@ private fun CampfireDockedSearchBar(
       onGenreClick = { genre -> eventSink(SearchUiEvent.OnGenreClick(genre)) },
     )
   }
-}
-
-@Composable
-private fun SearchResultContent(
-  query: String,
-  results: SearchResult,
-  onBookClick: (LibraryItem) -> Unit,
-  onAuthorClick: (Author) -> Unit,
-  onSeriesClick: (Series) -> Unit,
-  onTagClick: (BasicSearchResult) -> Unit,
-  onGenreClick: (BasicSearchResult) -> Unit,
-  onNarratorClick: (BasicSearchResult) -> Unit,
-  modifier: Modifier = Modifier,
-) {
-  when (results) {
-    SearchResult.Error -> EmptyState(stringResource(Res.string.search_results_error_message), modifier)
-    SearchResult.Loading -> LoadingContent(modifier)
-    is SearchResult.Success -> if (results.isEmpty && query.isNotBlank()) {
-      EmptyState(
-        message = {
-          Text(
-            text = buildAnnotatedString {
-              append(stringResource(Res.string.search_results_empty_message))
-              append(" ")
-              withStyle(SpanStyle(fontWeight = FontWeight.Bold)) {
-                append("\"")
-                append(query)
-                append("\"")
-              }
-            },
-          )
-        },
-        modifier = modifier,
-      )
-    } else if (results.isEmpty && query.isBlank()) {
-      EmptyState(
-        message = stringResource(Res.string.search_empty_message),
-        modifier = modifier,
-      )
-    } else {
-      CompositionLocalProvider(
-        LocalContentLayout provides ContentLayout.Root,
-      ) {
-        SuccessContent(
-          result = results,
-          onBookClick = onBookClick,
-          onNarratorClick = onNarratorClick,
-          onAuthorClick = onAuthorClick,
-          onSeriesClick = onSeriesClick,
-          onTagClick = onTagClick,
-          onGenreClick = onGenreClick,
-          modifier = modifier,
-        )
-      }
-    }
-  }
-}
-
-@Composable
-private fun LoadingContent(
-  modifier: Modifier = Modifier,
-) {
-  Box(
-    modifier = modifier.fillMaxSize(),
-    contentAlignment = Alignment.Center,
-  ) {
-    CircularProgressIndicator()
-  }
-}
-
-@OptIn(ExperimentalLayoutApi::class)
-@Composable
-private fun SuccessContent(
-  result: SearchResult.Success,
-  onBookClick: (LibraryItem) -> Unit,
-  onAuthorClick: (Author) -> Unit,
-  onSeriesClick: (Series) -> Unit,
-  onTagClick: (BasicSearchResult) -> Unit,
-  onGenreClick: (BasicSearchResult) -> Unit,
-  onNarratorClick: (BasicSearchResult) -> Unit,
-  modifier: Modifier = Modifier,
-) {
-  LazyVerticalGrid(
-    columns = GridCells.Adaptive(100.dp),
-    modifier = modifier,
-    horizontalArrangement = Arrangement.spacedBy(8.dp),
-    verticalArrangement = Arrangement.spacedBy(8.dp),
-    contentPadding = PaddingValues(
-      horizontal = 16.dp,
-      vertical = 8.dp,
-    ),
-  ) {
-    // Books
-    result.books.ifNotEmpty { books ->
-      headerItem("books") { stringResource(Res.string.header_books) }
-      items(
-        items = books,
-        key = { it.id },
-      ) { book ->
-        LibraryItemCard(
-          item = book,
-          modifier = Modifier
-            .animateItem()
-            .clickable {
-              onBookClick(book)
-            },
-        )
-      }
-    }
-
-    // Authors
-    result.authors.ifNotEmpty { authors ->
-      headerItem("authors") { stringResource(Res.string.header_authors) }
-      items(
-        items = authors,
-        key = { it.id },
-      ) { author ->
-        AuthorCard(
-          author = author,
-          modifier = Modifier
-            .animateItem()
-            .clickable {
-              onAuthorClick(author)
-            },
-        )
-      }
-    }
-
-    // Series
-    result.series.ifNotEmpty { series ->
-      headerItem("series") { stringResource(Res.string.header_series) }
-      items(
-        items = series,
-        key = { it.id },
-        span = { GridItemSpan(2) },
-      ) { s ->
-        ItemCollectionCard(
-          name = s.name,
-          description = s.description,
-          items = s.books ?: emptyList(),
-          itemSize = 100.dp,
-          modifier = Modifier
-            .fillMaxWidth()
-            .animateItem()
-            .clickable {
-              onSeriesClick(s)
-            },
-        )
-      }
-    }
-
-    // Narrators
-    result.narrators.ifNotEmpty { narrators ->
-      headerItem("narrators") { stringResource(Res.string.header_narrators) }
-      item(
-        span = { GridItemSpan(maxLineSpan) },
-      ) {
-        FlowRow(
-          modifier = modifier,
-          horizontalArrangement = Arrangement.spacedBy(8.dp),
-          verticalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-          narrators.forEach { narrator ->
-            BasicSearchResultChip(
-              result = narrator,
-              modifier = Modifier.clickable {
-                onNarratorClick(narrator)
-              },
-            )
-          }
-        }
-      }
-    }
-
-    // Tags
-    result.tags.ifNotEmpty { tags ->
-      headerItem("tags") { stringResource(Res.string.header_tags) }
-      item(
-        span = { GridItemSpan(maxLineSpan) },
-      ) {
-        FlowRow(
-          modifier = modifier.animateItem(),
-          horizontalArrangement = Arrangement.spacedBy(8.dp),
-          verticalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-          tags.forEach { tag ->
-            BasicSearchResultChip(
-              result = tag,
-              modifier = Modifier.clickable {
-                onTagClick(tag)
-              },
-            )
-          }
-        }
-      }
-    }
-
-    // Narrators
-    result.genres.ifNotEmpty { genres ->
-      headerItem("genres") { stringResource(Res.string.header_genres) }
-      item(
-        span = { GridItemSpan(maxLineSpan) },
-      ) {
-        FlowRow(
-          modifier = modifier,
-          horizontalArrangement = Arrangement.spacedBy(8.dp),
-          verticalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-          genres.forEach { genre ->
-            BasicSearchResultChip(
-              result = genre,
-              modifier = Modifier.clickable {
-                onGenreClick(genre)
-              },
-            )
-          }
-        }
-      }
-    }
-
-    item { Spacer(Modifier.height(16.dp)) }
-  }
-}
-
-fun LazyGridScope.headerItem(key: Any?, title: @Composable () -> String) {
-  item(
-    key = key,
-    span = { GridItemSpan(maxLineSpan) },
-    contentType = "header",
-  ) {
-    Text(
-      text = title(),
-      style = MaterialTheme.typography.labelLarge,
-      fontWeight = FontWeight.SemiBold,
-      modifier = Modifier
-        .padding(
-          vertical = 16.dp,
-        )
-        .animateItem(),
-    )
-  }
-}
-
-@OptIn(ExperimentalContracts::class)
-private inline fun <T> List<T>.ifNotEmpty(block: (List<T>) -> Unit) {
-  contract {
-    callsInPlace(block, InvocationKind.AT_MOST_ONCE)
-  }
-  if (isNotEmpty()) block(this)
 }
