+++ features/search/ui/src/commonMain/kotlin/app/campfire/search/ui/composables/BasicSearchResultChip.kt
@@ -33,6 +33,7 @@ internal fun BasicSearchResultChip(
     Row(
       verticalAlignment = Alignment.CenterVertically,
     ) {
+      Spacer(Modifier.width(8.dp))
       Text(result.name)
       Spacer(Modifier.width(8.dp))
       Text(
