+++ features/search/ui/src/commonMain/kotlin/app/campfire/search/ui/di/SearchUiComponent.kt
@@ -1,0 +1,10 @@
+package app.campfire.search.ui.di
+
+import app.campfire.core.di.UserScope
+import app.campfire.search.ui.SearchPresenterFactory
+import com.r0adkll.kimchi.annotations.ContributesTo
+
+@ContributesTo(UserScope::class)
+interface SearchUiComponent {
+  val searchPresenterFactory: SearchPresenterFactory
+}
